+++ packages/components/package.json
@@ -82,9 +82,7 @@
         "date-fns": "^3.6.0",
         "downshift": "^9.0.13",
         "react-colorful": "^5.6.1",
-        "react-day-picker": "^9.13.1",
-        "tailwind-merge": "^2.6.0",
-        "tailwind-variants": "^0.3.1"
+        "react-day-picker": "^9.13.1"
     },
     "devDependencies": {
         "@babel/core": "^7.29.0",
