+++ packages/components/src/helpers/screens.ts
@@ -1,0 +1,10 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const screens = {
+    xs: '390px',
+    sm: '600px',
+    md: '768px',
+    lg: '1024px',
+    xl: '1280px',
+    '2xl': '1536px',
+};
