+++ packages/components/tailwind.config.ts
@@ -3,14 +3,7 @@
 import frontifyTailwindConfig from '@frontify/fondue-tokens/tailwind';
 import { type Config } from 'tailwindcss';
 
-export const screens = {
-    xs: '390px',
-    sm: '600px',
-    md: '768px',
-    lg: '1024px',
-    xl: '1280px',
-    '2xl': '1536px',
-};
+import { screens } from './src/helpers/screens';
 
 export default {
     prefix: 'tw-',
